практика по GET и POST запросам, postman, и сервер на express
