+++ README.md
@@ -1,5 +1,7 @@
 # dp-todo-app
 
-Добавил функцию конструктор Task, переписал функцию createTask с использованием конструктора, и функции changeStatus, changePriority с использованием методов функции-конструктора Task. А так же доделал то, что еще не было сделано: Разбил программу на модули, добавил импорт задач из .json файла, добавил функцию todoListCreate которая создает массив туду-листа из импортируемого массива из json.
+Используя сервер на express и postman для запросов в качестве клиента - практика по GET и POST запросам. По роуту http://localhost:3000/tasks GET выдаёт список текущих задач из массива. POST добавляет новую задачу в массив задач, задача берется из json из body POST запроса.
 
-todo_v5_constructor_methods.js - главный файл
+За основу взял свою TODO APP немного переделав под текущую задачу.
+
+todo_v6-express.js - главный файл
