+++ functions/helpFunctions.js
