+++ todoList.json
@@ -1,11 +1,5 @@
 [
-    { "name": "сделать чай", "status": "In progress", "priority": "low" },
-    {
-        "name": "написать код тудушки",
-        "status": "In progress",
-        "priority": "low"
-    },
-    { "name": "поработать работу", "status": "Done", "priority": "high" },
-    { "name": "пойти на пробежку", "status": "ToDo", "priority": "high" },
-    { "name": "перестать много жрать", "status": "ToDo", "priority": "medium" }
+    { "name": "сделать чай", "status": "ToDo", "priority": "low" },
+
+    { "name": "поработать работу", "status": "Done", "priority": "high" }
 ]
